Providers are disabled unless specifically configured
This is so that doing `models.refresh!` won't attempt to use every
provider API, resulting in errors unless a valid key is given for every one.

If this is too disruptive to be default behavior, then this could be a
"default to offline" mode that needs to be explicitly turned on perhaps
with an environment variable; in its absence, behavior would be the same
as before.

This is all so that `models.refresh!` can be called freely to populate
Ollama models at runtime.