Add OpenAI URL override and custom req headers
This commit introduces the configuration primitives necessary for
semantic caching proxy services for OpenAI requests.Specifically,
the ability to override the OpenAI root API URL path and the
ability to add custom headers to the request payload in the form
of a hash.

This commit does not introduce tests, and should not be merged
until a subsequent commit lands with proper testing.