Fix unsafe Performance/UnfreezeString offenses
While `String#dup` is as fast as `+''` in Ruby >= 3.3, this is not the
case for `String.new`, so this cop has been applied only in production
code.

RuboCop considers this change unsafe because `+''` returns an UTF-8
string, while `String.new` returns an ASCII string.

In this context, this should not be a breaking change.

```
rubocop --only Performance/UnfreezeString -A
```

A manual fix has been applied in ActsAs concern to use empty string
literal for ActiveRecord `content` assignment instead of `String.new` or
`+''`. Passing `''` directly to ActiveRecord is sufficient since the
string isn't mutated.

Bench:

```
ruby 3.4.5 (2025-07-16 revision 20cda200d3) +PRISM [arm64-darwin24]

Comparison (IPS):
                 +'': 15938643.5 i/s
          String.new: 11016752.0 i/s - 1.45x  (± 0.00) slower
```

Ref:
- https://docs.rubocop.org/rubocop-performance/cops_performance.html#performanceunfreezestring
- https://github.com/fastruby/fast-ruby?tab=readme-ov-file#stringdup-vs-string-code