Preserve models from unconfigured providers
This fixes an issue where models from unconfigured
providers are not preserved when running rake models:update.

Since calling refresh! refetches the list of models and fully
replaces the old list, it will also completely skip unconfigured
providers in the new list (except if they're coming from parsera
as well).

This change ensures that configured providers are fetched fresh and
unconfigured ones are preserved as they are, allow for partial
model updating.